[Querybuilder] Fix
There was a bug in the multi-select. If you did col_x in ['a', 'col with space'] it was also putting tickmarks for the column names within the brackets.

We went back to the original approach of looping through an array of column names. Only in this case, we considered only strings that explicitly end in a character and not a space (which indicates another string).

Related to https://github.com/molinsp/eigendata/commit/ecac005b1e4cad5c4a36d53406f17b9728ce2e05